requirements.txt
customtkinter
pyautogui
pystray
pillow
pyinstaller
